requirements.txt
web3~=6.14
azure-identity~=1.16
azure-keyvault-secrets~=4.9
python-dotenv~=1.0
fastapi~=0.110
uvicorn[standard]~=0.24
